+++ KSR/Tools/Features/KeyWordsFirstParagraphArticleBodyFeature.cs
@@ -9,7 +9,7 @@ namespace KSR.Tools.Features
     {
         public ISimilarityFunction SimilarityFunction { get; set; }
 
-        public decimal Calc(Article article, List<string> keyWords)
+        public double Calc(Article article, List<string> keyWords)
         {
             var count = 0;
             var words = article.Paragraphs[0];
