+++ KSR/Tools/Helpers/DictHelper.cs
@@ -4,10 +4,10 @@ namespace KSR.Tools.Helpers
 {
     public static class DictHelper
     {
-        public static string DictionaryToString(Dictionary<string, decimal> dictionary)
+        public static string DictionaryToString(Dictionary<string, double> dictionary)
         {
             string dictionaryString = "{";
-            foreach (KeyValuePair<string, decimal> keyValues in dictionary)
+            foreach (KeyValuePair<string, double> keyValues in dictionary)
             {
                 dictionaryString += keyValues.Key + " : " + keyValues.Value + ", ";
             }
