+++ KSR/Tools/SimliarityFunctions/LevenshteinFunction.cs
@@ -3,7 +3,7 @@ namespace KSR.Tools.SimliarityFunctions
 {
     public class LevenshteinFunction : ISimilarityFunction
     {
-        public decimal CalculateSimilarity(string firstWord, string secondWord)
+        public double CalculateSimilarity(string firstWord, string secondWord)
         {
             return Fastenshtein.Levenshtein.Distance(firstWord, secondWord);
         }
